test(prompts): consolidate redundant acceptanceContext and scope tests
- Remove 2 packed mega-tests from builder.test.ts that duplicated
  coverage already in builder-acceptance.test.ts (US-001 AC4/AC5)
- Migrate their 4 unique assertions (fenced code block, multiple
  entries, determinism, no-truncation marker) as focused named tests
  into builder-acceptance.test.ts
- Delete rectifier-builder-scope.test.ts (single-test file); absorb
  the scope-guidance test into builders/rectifier-builder.test.ts
  where reviewRectification is already tested extensively
