+++ test/unit/prompts/builder.test.ts
@@ -318,49 +318,3 @@ describe("PromptBuilder — batch role: build()", () => {
 });
 
 
-// ---------------------------------------------------------------------------
-// US-001 AC4: PromptBuilder.acceptanceContext() — stores entries, builds section
-// US-001 AC5: When acceptanceContext() not called, build() is identical to baseline
-// ---------------------------------------------------------------------------
-
-describe("PromptBuilder — acceptanceContext() method (US-001 AC4)", () => {
-  test(".acceptanceContext() chainable; content/path present; appears after story; fenced code block; multiple entries", async () => {
-    const story = makeStory({ title: "STORY_BEFORE_ACCEPTANCE" });
-    const entries = [{ testPath: "test/unit/my-feature.test.ts", content: "ACCEPTANCE_CONTENT_MARKER" }];
-    const builder = PromptBuilder.for("implementer").story(story).acceptanceContext(entries);
-    expect(builder).toBeInstanceOf(PromptBuilder);
-    const prompt = await builder.build();
-    expect(prompt).toContain("ACCEPTANCE_CONTENT_MARKER");
-    expect(prompt).toContain("test/unit/my-feature.test.ts");
-    expect(prompt.indexOf("STORY_BEFORE_ACCEPTANCE")).toBeLessThan(prompt.indexOf("ACCEPTANCE_CONTENT_MARKER"));
-    expect(prompt).toContain("```typescript");
-
-    const multiEntries = [
-      { testPath: "test/a.test.ts", content: "CONTENT_A" },
-      { testPath: "test/b.test.ts", content: "CONTENT_B" },
-    ];
-    const multiPrompt = await PromptBuilder.for("implementer").story(makeStory()).acceptanceContext(multiEntries).build();
-    expect(multiPrompt).toContain("test/a.test.ts");
-    expect(multiPrompt).toContain("test/b.test.ts");
-    expect(multiPrompt).toContain("CONTENT_A");
-    expect(multiPrompt).toContain("CONTENT_B");
-  });
-});
-
-describe("PromptBuilder — no acceptance section when acceptanceContext() not called (US-001 AC5)", () => {
-  test("build() without acceptanceContext(): no truncated marker, deterministic, no acceptance-only markers", async () => {
-    const story = makeStory({ title: "BASELINE_STORY_AC5" });
-    const promptA = await PromptBuilder.for("implementer").story(story).build();
-    expect(promptA).not.toContain("[truncated — full file at");
-    const promptB = await PromptBuilder.for("implementer").story(story).build();
-    expect(promptA).toBe(promptB);
-
-    const withAcceptance = await PromptBuilder.for("implementer")
-      .story(makeStory())
-      .acceptanceContext([{ testPath: "test/x.test.ts", content: "UNIQUE_AC5_MARKER" }])
-      .build();
-    const withoutAcceptance = await PromptBuilder.for("implementer").story(makeStory()).build();
-    expect(withAcceptance).toContain("UNIQUE_AC5_MARKER");
-    expect(withoutAcceptance).not.toContain("UNIQUE_AC5_MARKER");
-  });
-});
