+++ test/unit/prompts/builders/rectifier-builder.test.ts
@@ -12,6 +12,7 @@
  */
 
 import { describe, expect, test } from "bun:test";
+import { makeStory } from "../../../helpers";
 import { RectifierPromptBuilder } from "@/prompts";
 import type { ReviewCheckResult } from "@/review";
 import type { Finding } from "@/findings";
@@ -536,6 +537,31 @@ describe("RectifierPromptBuilder.reviewRectification — blocking-only defensive
     const reviewPrompt = RectifierPromptBuilder.reviewRectification(checks, makeStory());
     // The semantic path uses formatCheckErrors (raw output), not structured findings
     expect(reviewPrompt).toContain("Semantic review failed");
+  });
+});
+
+describe("RectifierPromptBuilder.reviewRectification — scope guidance (package-local prerequisite)", () => {
+  test("allows package-local prerequisite fixes before classifying a failure as sibling spillover", () => {
+    const story = makeStory({
+      id: "US-001",
+      title: "Prefer const over let in greet()",
+      acceptanceCriteria: ["bun run typecheck exits with code 0"],
+      routing: { testStrategy: "no-test", complexity: "simple", reasoning: "no-op change" },
+    });
+    const failedChecks: ReviewCheckResult[] = [
+      {
+        check: "typecheck",
+        success: false,
+        command: "bun run typecheck",
+        exitCode: 2,
+        output: "error TS2688: Cannot find type definition file for 'bun-types'.",
+        durationMs: 1,
+      },
+    ];
+    const prompt = RectifierPromptBuilder.reviewRectification(failedChecks, story);
+    expect(prompt).toContain("smallest package-local fix is required");
+    expect(prompt).toContain("TEST_EDIT_REASON: sibling_scope");
+    expect(prompt).not.toContain("When a lint or typecheck error is in a file you did NOT create or modify in this turn");
   });
 });
